bike update and delete not found error handling

+++ src/app/modules/bike/bike.services.ts
@@ -1,5 +1,7 @@
+import { AppError } from '../../errors/AppError';
 import { TBike } from './bike.interface';
 import { Bike } from './bike.model';
+import status from 'http-status';
 
 const createBike = async (payload: TBike) => {
   const data = await Bike.create(payload);
@@ -13,6 +15,13 @@ const retrieveAllBikes = async (query: any) => {
 };
 
 const updateBikes = async (id: string, payload: Partial<TBike>) => {
+  // check is bike exist?
+  const isBikeExist = await Bike.findById(id);
+
+  if (!isBikeExist) {
+    throw new AppError(status.NOT_FOUND, 'Bike is no found!');
+  }
+
   const data = await Bike.findByIdAndUpdate(id, payload, {
     new: true,
     runValidators: true,
@@ -21,6 +30,11 @@ const updateBikes = async (id: string, payload: Partial<TBike>) => {
 };
 
 const deleteBikes = async (id: string) => {
+  const isBikeExist = await Bike.findById(id);
+
+  if (!isBikeExist) {
+    throw new AppError(status.NOT_FOUND, 'Bike is no found!');
+  }
   const data = await Bike.findByIdAndDelete(id);
   return data;
 };
